Modified to return data byte by byte.
This is a specific hack intended for termshark and isn't really suitable for a
pull request. I want to be able to monitor a pcap file for changes and send
those changes to stdout without necessarily waiting for a newline. This is
inefficient because I send the data byte by byte, and could certaainly be
improved. But since tail has figured out EOF issues, I decided it's easier to
start here than write a binary tail program myself.